Remove `PieceProvider::get_piece_from_dsn_cache_with_retries()` as unnecessary
